+++ src/test/java/com/jinddung2/givemeticon/item/presentation/ItemControllerTest.java
@@ -2,11 +2,12 @@ package com.jinddung2.givemeticon.item.presentation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jinddung2.givemeticon.common.config.WebConfig;
-import com.jinddung2.givemeticon.common.interceptor.AuthInterceptor;
+import com.jinddung2.givemeticon.common.security.interceptor.AuthInterceptor;
 import com.jinddung2.givemeticon.domain.item.controller.ItemController;
 import com.jinddung2.givemeticon.domain.item.dto.request.ItemCreateRequest;
 import com.jinddung2.givemeticon.domain.item.facade.ItemCreationFacade;
-import com.jinddung2.givemeticon.user.application.LoginService;
+import com.jinddung2.givemeticon.domain.item.service.ItemService;
+import com.jinddung2.givemeticon.domain.user.service.LoginService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -37,6 +38,9 @@ class ItemControllerTest {
 
     @MockBean
     ItemCreationFacade itemCreationFacade;
+
+    @MockBean
+    ItemService itemService;
 
     ItemCreateRequest itemCreateRequest;
     int brandId = 10;
